backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
tiktoken==0.5.1
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6